1. 어플리케이션으로 제작하면 공증이 필요해서 RPM 으로 생성 후 스크립트로 실행하도록 수정 > 추후에 공증받아서 해야함 2. RPM 설치 후 sudo chown -R [groupname]:[username] urtsp_client 명령어 입력 필요 3. Config 파일 열어서 수정할 수 있도록 기능 추가

+++ src/main/java/com/rtsp/client/media/netty/module/base/RtspUnit.java
@@ -328,35 +328,39 @@ public class RtspUnit {
         }
 
         if (configManager.isDeleteMp4()) {
-            File mp4File = new File(mp4FilePath);
-            if (mp4File.exists()) {
-                if (mp4File.delete()) {
-                    logger.debug("({}) Success to delete the mp4 file. (path={})", rtspUnitId, mp4FilePath);
+            if (mp4FilePath != null) {
+                File mp4File = new File(mp4FilePath);
+                if (mp4File.exists()) {
+                    if (mp4File.delete()) {
+                        logger.debug("({}) Success to delete the mp4 file. (path={})", rtspUnitId, mp4FilePath);
+                    }
                 }
             }
         }
 
-        File tempRootPathDirectory = new File(tempFileRootPath);
-        if (configManager.isDeleteM3u8() && configManager.isDeleteTs() && configManager.isDeleteMp4()) {
-            if (tempRootPathDirectory.exists()) {
+        if (tempFileRootPath != null) {
+            File tempRootPathDirectory = new File(tempFileRootPath);
+            if (configManager.isDeleteM3u8() && configManager.isDeleteTs() && configManager.isDeleteMp4()) {
+                if (tempRootPathDirectory.exists()) {
+                    if (tempRootPathDirectory.delete()) {
+                        logger.debug("({}) The temp root directory is deleted. (path={})", rtspUnitId, tempFileRootPath);
+                    }
+                }
+            }
+
+            File[] deleteFolderList = tempRootPathDirectory.listFiles();
+            if (deleteFolderList != null) {
+                for (File file : deleteFolderList) {
+                    if (file.delete()) {
+                        logger.debug("({}) The temp file is deleted. (path={})", rtspUnitId, file.getAbsolutePath());
+                    }
+                }
+            }
+
+            if (deleteFolderList == null || deleteFolderList.length == 0) {
                 if (tempRootPathDirectory.delete()) {
                     logger.debug("({}) The temp root directory is deleted. (path={})", rtspUnitId, tempFileRootPath);
                 }
-            }
-        }
-
-        File[] deleteFolderList = tempRootPathDirectory.listFiles();
-        if (deleteFolderList != null) {
-            for (File file : deleteFolderList) {
-                if (file.delete()) {
-                    logger.debug("({}) The temp file is deleted. (path={})", rtspUnitId, file.getAbsolutePath());
-                }
-            }
-        }
-
-        if (deleteFolderList == null || deleteFolderList.length == 0) {
-            if (tempRootPathDirectory.delete()) {
-                logger.debug("({}) The temp root directory is deleted. (path={})", rtspUnitId, tempFileRootPath);
             }
         }
     }
